requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
langchain==0.3.26
# langchain-openai==0.0.2  # 注释掉OpenAI依赖
langchain-community==0.3.27  # 添加社区包支持
dashscope>=1.13.3  # 添加通义千问依赖
pydantic>=2.4.2
pydantic-settings==2.10.1
python-dotenv>=1.0.0
sqlalchemy==2.0.23
python-jose>=3.3.0
passlib>=1.7.4
python-multipart==0.0.6
email-validator==2.2.0
httpx>=0.25.1
tenacity>=8.2.3 